Serve construction-v1 Addresses from promoted slices in v2 forward (#193)
Route /v2/forward structured lookups through a promoted
slice-YYYY-MM-DD.N/families/addresses/ tree when the release's addresses
family declares the OAV1ART construction format, mirroring the #191
Places lane:

- validate_family accepts OAV1ART for family_slice sources only, with
  the routing.json entrypoint under the shared 8 MiB cap and the
  address-transform-v1 normalization identity; address-reduce-2 keeps
  its exact contract and every mixed/unknown combination fails closed.
- Admission authenticates the family manifest against the release,
  proves routing.json byte-for-byte against the release-pinned
  entrypoint identity, validates the per-country inclusive route-hash
  envelope table (sorted, overlap-free, content-addressed .av1 names),
  requires every routed object to be attested, and stream-verifies a
  deterministic sample of the up-to-three smallest routed objects under
  a 32 MiB per-object bound. Planet artifacts approach the 2 GiB cap,
  so larger objects stay manifest-pinned and structurally self-checked
  at serving time.
- Serving resolves the normalized key's route hash to its envelope
  (country miss = out_of_coverage, inter-envelope gap = provably empty)
  and answers from the single owning OAV1ART object through a bounded
  range-read planner: one 44-byte header, <= 27 logarithmic 24-byte
  index probes (64 hard cap), one <= 12,312-byte candidate window, and
  one <= 8 MiB payload run, never whole-object hydration. The
  whole-buffer decoder stays as the reference implementation and native
  tests cross-check the planner against it on identical bytes,
  including the encoder-emitted fixture.
- AddressPageRecord widens source_row_index to u64 (reduce-2 decode
  still enforces its u32 bound) so both lanes emit one feature shape.

+++ crates/geocoder-worker/src/address_pages.rs
@@ -46,7 +46,10 @@ pub(crate) struct AddressPageRecord {
     pub latitude: f64,
     pub source_object_index: u32,
     pub source_row_group: u32,
-    pub source_row_index: u32,
+    /// `u64` so the construction-v1 lane (which stores 64-bit row indexes)
+    /// shares this record shape; the reduce-2 page decoder still enforces its
+    /// own u32 bound before widening.
+    pub source_row_index: u64,
     pub country: String,
     pub postal_city: String,
     pub postcode: String,
@@ -245,8 +248,10 @@ fn decode_useful_page_measured(bytes: &[u8]) -> Result<(Vec<AddressPageRecord>, 
             .map_err(|_| PageError::new("source object index is too large"))?;
         let source_row_group = u32::try_from(reader.uvarint()?)
             .map_err(|_| PageError::new("source row group is too large"))?;
-        let source_row_index = u32::try_from(reader.uvarint()?)
-            .map_err(|_| PageError::new("source row index is too large"))?;
+        let source_row_index = u64::from(
+            u32::try_from(reader.uvarint()?)
+                .map_err(|_| PageError::new("source row index is too large"))?,
+        );
         let mut display = [0_usize; 6];
         for slot in display.iter_mut() {
             *slot = dictionary_id(&mut reader, strings.len())?;
